Figured out mocks and got a couple tests up and running
