--- src/test/kotlin/com/merricklabs/portlandbus/mocks/TestMock.kt
@@ -1,5 +1,0 @@
-package com.merricklabs.portlandbus.mocks
-
-interface TestMock {
-    fun reset()
-}
